# docs/swagger.yaml
basePath: /
definitions:
...
  routes.CreateCryptoPaymentRequest:
    properties:
      plan:
        example: pro
        type: string
    type: object
...
  routes.PaystackInitializeRequest:
    properties:
      callback:
        example: https://app.ganium.ai/payment/callback
        type: string
      callback_url:
        example: https://app.ganium.ai/payment/callback
        type: string
      callbackUrl:
        example: https://app.ganium.ai/payment/callback
        type: string
      mode:
        example: hosted
        type: string
      plan:
        example: pro
        type: string
      plan_code:
        example: pro
        type: string
      planCode:
        example: pro
        type: string
      planName:
        example: Pro
        type: string
      reference:
        example: GANIUM-123456
        type: string
    type: object
  routes.PaystackVerifyRequest:
    properties:
      reference:
        example: GANIUM-123456
        type: string
    type: object
...
  routes.SubmitCryptoPaymentRequest:
    properties:
      payment_id:
        example: GAN-SJLY-123456789
        type: string
      signature:
        example: 5abc123...
        type: string
      user_wallet:
        example: 7abc123...
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ganium.ai
    name: Ganium Support
  description: Ganium AI scam detection, security investigation, scanning, authentication
    and payment API.
  license:
    name: Proprietary
  termsOfService: https://ganium.ai/terms
  title: Ganium API
  version: "1.0"
paths:
...
  /api/crypto/payment/create:
    post:
      consumes:
      - application/json
      description: Creates a pending SJLY payment intent.
      parameters:
      - description: Crypto payment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.CreateCryptoPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create SJLY payment
      tags:
      - Crypto Payments
  /api/crypto/payment/status:
    get:
      description: Returns the current status of an SJLY payment.
      parameters:
      - description: Crypto payment ID
        in: query
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get crypto payment status
      tags:
      - Crypto Payments
  /api/crypto/payment/submit:
    post:
      consumes:
      - application/json
      description: Saves the Solana transaction signature for verification.
      parameters:
      - description: Transaction
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.SubmitCryptoPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit SJLY transaction
      tags:
      - Crypto Payments
...
  /api/paystack/callback:
    get:
      description: Handles Paystack payment callback.
      parameters:
      - description: Paystack transaction reference
        in: query
        name: reference
        required: true
        type: string
      - description: Payment status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Paystack callback
      tags:
      - Paystack
  /api/paystack/initialize:
    post:
      consumes:
      - application/json
      description: Initializes a Paystack payment transaction.
      parameters:
      - description: Payment initialization
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.PaystackInitializeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initialize Paystack payment
      tags:
      - Paystack
  /api/paystack/plans:
    get:
      description: Returns available Ganium subscription plans.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get payment plans
      tags:
      - Paystack
  /api/paystack/stream:
    get:
      description: Streams Paystack payment status updates using SSE.
      parameters:
      - description: Paystack payment reference
        in: query
        name: reference
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Stream payment status
      tags:
      - Paystack
  /api/paystack/verify:
    post:
      consumes:
      - application/json
      description: Verifies a Paystack transaction and credits the user's account.
      parameters:
      - description: Payment verification request
        in: body
        name: request
        schema:
          $ref: '#/definitions/routes.PaystackVerifyRequest'
      - description: Payment reference
        in: query
        name: reference
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verify Paystack payment
      tags:
      - Paystack
...
  /paystack/webhook:
    post:
      consumes:
      - application/json
      description: Receives and verifies Paystack webhook events.
      parameters:
      - description: Paystack webhook signature
        in: header
        name: x-paystack-signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Paystack webhook
      tags:
      - Paystack
...
securityDefinitions:
  BearerAuth:
    description: 'JWT authentication. Enter your token using the format: Bearer <token>.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"